Criando uma classe genérica para testes
